fixed world preview compat + other changes

+++ common/src/main/java/raccoonman/reterraforged/world/worldgen/continent/simple/SimpleRiverGenerator.java
@@ -40,7 +40,7 @@ public class SimpleRiverGenerator extends BaseRiverGenerator<SimpleContinent> {
 			float x3 = x + dx * length;
 			float z3 = z + dz * length;
 			float valleyWidth = 275.0F * River.MAIN_VALLEY.next(random);
-			River river = new River((float) (int) x2, (float) (int) z2, (float) (int) x3, (float) (int) z3);
+			River river = new River((int) x2, (int) z2, (int) x3, (int) z3);
 			RiverPopulator.Settings settings = BaseRiverGenerator.creatSettings(random);
 			settings.fadeIn = this.main.fade;
 			settings.valleySize = valleyWidth;
